92-legal-case-researcher/requirements.txt
requests>=2.28.0
pyyaml>=6.0
click>=8.1.0
rich>=13.0.0
streamlit>=1.28.0
fastapi>=0.104.0
uvicorn>=0.24.0
pydantic>=2.0.0
pytest>=7.4.0
